requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==23.1.0
billiard==4.1.0
brotlipy==0.7.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
Django==4.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-yasg==1.20.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.2
MarkupSafe==2.1.1
packaging==23.1
pip==23.3
prompt-toolkit==3.0.39
pycparser==2.21
pyOpenSSL==23.2.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==5.4.1
redis==5.0.1
requests==2.31.0
ruamel.yaml==0.17.21
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.8
wheel==0.41.2
